requirements.txt
eyed3>=0.9.6
PySide6>=6.5.0
openai>=1.0.0
python-dotenv>=1.0.0
pygame>=2.5.0 